Added special offers and updated ordering logic
Added mechanics for viewing orders from the archive with navigation

Added admin db class with shop settings

+++ src/handlers/payment_processing/from_balance.py
@@ -1,5 +1,6 @@
 import datetime
 from random import randint
+from typing import List
 
 from aiogram import F, types
 from aiogram.fsm.context import FSMContext
@@ -32,11 +33,8 @@ from utils.states import NewOrder, Payment
 # url
 # t.me/...
 
-async def create_order(key: StorageKey, data: dict):
+async def create_order(key: StorageKey, service_id, quantity, url: str, data: dict, profit=0):
     user_id = key.user_id
-    quantity = data['quantity']
-    url = data['url']
-    service_id: str = data['service_id']
     total_amount = data['total_amount']
 
     platform = users.get_user_platform(user_id)
@@ -52,11 +50,23 @@ async def create_order(key: StorageKey, data: dict):
         'total_amount': total_amount
     }
 
+    if profit:
+        order_info['profit'] = profit
+
     order_id = await api.create_new_order(user_id, service_id, url, quantity)
     print(f'Оформил заказ номер {order_id}')
     orders.new_order(user_id, platform, order_id, order_info)
 
     return order_id
+
+
+async def create_multiple_order(key: StorageKey, services_and_amount: List[tuple], url: str, data: dict) -> List[int]:
+    order_ids = []
+    for service_id, quantity in services_and_amount:
+        order_id = await create_order(key, service_id, quantity, url, data)
+        order_ids.append(order_id)
+
+    return order_ids
 
 
 @dp.callback_query(F.data == 'payment_method_internal_balance', Payment.choosing_method)
@@ -66,6 +76,8 @@ async def _(query: types.CallbackQuery, state: FSMContext):
     key = StorageKey(bot.id, chat_id, user_id)
     lang = users.get_user_lang(user_id)
     data = await storage.get_data(key)
+    url = data['url']
+    hot_order = data['hot_order']
     total_amount: float = data['total_amount']
     user_balance = users.get_balance(user_id)
     currency = 'RUB'
@@ -73,12 +85,23 @@ async def _(query: types.CallbackQuery, state: FSMContext):
     await query.answer()
 
     if user_balance >= total_amount:
-        order_id = await create_order(key, data)
         current_balance = user_balance - total_amount
+        if not hot_order:
+            quantity = data['quantity']
+            service_id = data['service_id']
+            profit = data['profit']
+            order_id = await create_order(key, service_id, quantity, url, data, profit)
+            message = messages.order_is_created[lang].format(order_id=order_id, currency=currency,
+                                                             total_amount=total_amount, current_balance=current_balance)
+            await query.message.delete()
+
+        else:
+            services_and_amount: List[tuple] = data['services_and_amount']
+            order_ids: List[int] = await create_multiple_order(key, services_and_amount, url, data)
+            _ids = ', '.join([str(i) for i in order_ids])
+            message = messages.order_is_created[lang].format(order_ids=order_ids, currency=currency,
+                                                             total_amount=total_amount, current_balance=current_balance)
         users.update_balance(user_id, current_balance)
-        message = messages.order_is_created[lang].format(order_id=order_id, currency=currency,
-                                                         total_amount=total_amount, current_balance=current_balance)
-        await query.message.delete()
 
     else:
         message = messages.not_enough_money[lang].format(current_balance=user_balance, currency=currency)
